requirements.txt
Brlapi==0.6.5
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.7.1
cupshelpers==1.0
decorator==4.0.11
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
entrypoints==0.2.2.post2
httplib2==0.9.2
idna==2.7
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itypes==1.1.0
Jinja2==2.10
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.1
kazam==1.4.5
keyring==10.1
keyrings.alt==1.3
louis==3.0.0
Mako==1.0.6
MarkupSafe==1.0
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.2.0
notebook==4.2.3
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.22.0
pysmbc==1.0.15.6
python-apt==1.4.0b3
python-debian==0.1.30
python-debianbts==2.6.1
python-xlib==0.14
pytz==2018.5
pyxdg==0.25
pyzmq==16.0.2
reportbug==7.1.7
reportlab==3.3.0
requests==2.19.1
SecretStorage==2.3.1
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.10.0
termcolor==1.1.0
terminado==0.6
tornado==4.4.3
traitlets==4.3.1
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
